backend/django/api/.history/requirements_20211005202307.txt
aiohttp==3.7.4.post0
aioredis==1.3.1
appdirs==1.4.4
argon2-cffi==21.1.0
asgiref==3.3.4
async-generator==1.10
async-timeout==3.0.1
attrs==20.3.0
Automat==20.2.0
backcall==0.2.0
backports.entry-points-selectable==1.1.0
beautifulsoup4==4.9.3
bleach==3.3.0
branca==0.4.2
bs4==0.0.1
certifi==2020.12.5
cffi==1.14.4
channels==3.0.4
channels-redis==3.3.1
chardet==4.0.0
charset-normalizer==2.0.4
chromedriver-autoinstaller==0.2.2
click==7.1.2
colorama==0.4.4
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==35.0.0
cycler==0.10.0
daphne==3.0.2
debugpy==1.4.3
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.1
dj-database-url==0.5.0
Django==3.2.1
django-bootstrap-v5==1.0.0
django-cors-headers==3.9.0
django-extensions==3.1.1
django-heroku==0.0.0
django-pandas==0.6.4
django-seed==0.2.2
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
drf-yasg==1.20.0
entrypoints==0.3
et-xmlfile==1.1.0
Faker==8.14.1
filelock==3.0.12
Flask==1.1.2
flex==6.14.1
folium==0.12.1
greenlet==1.1.1
gunicorn==20.1.0
haversine==2.5.1
hiredis==2.0.0
hyperlink==21.0.0
idna==2.10
importlib-metadata==2.1.1
incremental==21.3.0
inflection==0.5.1
ipykernel==6.4.1
ipython==7.28.0
ipython-genutils==0.2.0
itsdangerous==2.0.1
itypes==1.2.0
jedi==0.18.0
Jinja2==3.0.1
joblib==1.0.1
jsonpointer==2.1
jsonschema==4.0.1
jupyter-client==7.0.5
jupyter-core==4.8.1
jupyterlab-pygments==0.1.2
kiwisolver==1.3.2
MarkupSafe==2.0.1
matplotlib==3.4.3
matplotlib-inline==0.1.3
mistune==0.8.4
msgpack==1.0.2
multidict==5.1.0
mysql-connector-python==8.0.25
mysqlclient==2.0.3
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.5.1
notebook==6.2.0
numpy==1.20.3
openpyxl==3.0.8
packaging==20.9
pandas==1.2.4
pandocfilters==1.4.3
parso==0.8.1
pbr==5.6.0
pickleshare==0.7.5
Pillow==8.1.2
platformdirs==2.2.0
plotly==5.3.1
prometheus-client==0.9.0
prompt-toolkit==3.0.18
protobuf==3.18.0
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.10.0
PyJWT==1.7.1
PyMySQL==1.0.2
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyperclip==1.8.2
pyrsistent==0.18.0
python-dateutil==2.8.1
pytz==2021.1
pywin32==301
pywinpty==1.1.4
PyYAML==5.4.1
pyzmq==22.3.0
requests==2.25.1
rfc3987==1.3.8
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
scikit-learn==0.24.2
scipy==1.7.1
seaborn==0.11.2
selenium==3.141.0
semantic-version==2.8.5
Send2Trash==1.5.0
service-identity==21.1.0
setuptools-rust==0.11.6
simplejson==3.17.5
six==1.15.0
sklearn==0.0
soupsieve==2.2
SQLAlchemy==1.4.25
sqlparse==0.4.1
stack-require==0.0.2.1
stack-scaffold==0.0.2b1
stevedore==3.4.0
strict-rfc3339==0.7
template-render==0.1.0
tenacity==8.0.1
terminado==0.9.2
testpath==0.4.4
text-unidecode==1.3
threadpoolctl==2.2.0
toml==0.10.2
tornado==6.1
traitlets==5.0.5
Twisted==21.7.0
twisted-iocpsupport==1.0.2
txaio==21.2.1
typing-extensions==3.10.0.2
uritemplate==3.0.1
urllib3==1.26.3
validate-email==1.3
virtualenv==20.4.2
virtualenv-clone==0.5.6
virtualenvwrapper==4.8.4
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
whitenoise==5.2.0
yarl==1.6.3
zipp==3.4.1
zope.interface==5.4.0
